requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.2.0
braceexpand==0.1.7
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
discord.py==2.0.0a3575+g45d498c1
idna==3.3
import-expression==1.1.4
jishaku==2.3.2
multidict==5.2.0
requests==2.26.0
six==1.16.0
typing-extensions==4.0.0
update-checker==0.18.0
urllib3==1.26.7
wavelink==1.0.0b30
yarl==1.7.2
